:hammer: first steps to split ui
